parcel logik 1

+++ app/src/main/java/com/novatec/epitak/util/SizeUtils.kt
@@ -2,19 +2,18 @@ package com.novatec.epitak.util
 
 import android.content.Context
 import android.util.DisplayMetrics
-import kotlin.math.roundToInt
 
 
 object SizeUtils {
 
-    fun dpToPx(context: Context, dp: Int): Float {
+    fun dpToPx(context: Context, dp: Int): Int {
         val displayMetrics: DisplayMetrics = context.resources.displayMetrics
-        return (dp * (displayMetrics.xdpi / DisplayMetrics.DENSITY_DEFAULT))
+        return (dp * (displayMetrics.xdpi / DisplayMetrics.DENSITY_DEFAULT)).toInt()
     }
 
-    fun pxToDp(context: Context, px: Int): Float {
+    fun pxToDp(context: Context, px: Int): Int {
         val displayMetrics: DisplayMetrics = context.resources.displayMetrics
-        return (px / (displayMetrics.xdpi / DisplayMetrics.DENSITY_DEFAULT))
+        return (px / (displayMetrics.xdpi / DisplayMetrics.DENSITY_DEFAULT)).toInt()
     }
 
 }
